+++ cmd/app/options/options.go
@@ -28,9 +28,9 @@ import (
 	"gorm.io/driver/mysql"
 	"gorm.io/gorm"
 	"k8s.io/client-go/kubernetes"
+	"k8s.io/client-go/tools/clientcmd"
 
 	"github.com/caoyingjunz/gopixiu/cmd/app/config"
-	"github.com/caoyingjunz/gopixiu/pkg/core"
 	"github.com/caoyingjunz/gopixiu/pkg/db"
 	"github.com/caoyingjunz/gopixiu/pkg/log"
 	"github.com/caoyingjunz/gopixiu/pkg/types"
@@ -56,10 +56,11 @@ type Options struct {
 	CicdDriver *gojenkins.Jenkins
 
 	// ClientSets
-	ClientSets map[string]*kubernetes.Clientset
+	ClientSets *kubernetes.Clientset
 
 	// ConfigFile is the location of the pixiu server's configuration file.
 	ConfigFile string
+	KubeConfig string // Path to a kubeConfig.
 }
 
 func NewOptions() (*Options, error) {
@@ -100,6 +101,7 @@ func (o *Options) Complete() error {
 // BindFlags binds the pixiu Configuration struct fields
 func (o *Options) BindFlags(cmd *cobra.Command) {
 	cmd.Flags().StringVar(&o.ConfigFile, "configfile", "", "The location of the gopixiu configuration file")
+	cmd.Flags().StringVar(&o.KubeConfig, "kubeconfig", "", "The location of the kubeconfig file") // TODO: will be removed
 }
 
 func (o *Options) register() error {
@@ -156,9 +158,12 @@ func (o *Options) registerCicdDriver() error {
 }
 
 func (o *Options) registerClientSets() error {
-	var err error
-
-	o.ClientSets, err = core.RegisterClientSets(o.Factory)
+	kubeConfig, err := clientcmd.BuildConfigFromFlags("", o.KubeConfig)
+	if err != nil {
+		return err
+	}
+	// 生成clientSet
+	o.ClientSets, err = kubernetes.NewForConfig(kubeConfig)
 	if err != nil {
 		return err
 	}
