+++ cmd/app/server.go
@@ -26,8 +26,8 @@ import (
 	"k8s.io/klog/v2"
 
 	"github.com/caoyingjunz/gopixiu/api/server/router/cicd"
+	"github.com/caoyingjunz/gopixiu/api/server/router/cloud"
 	"github.com/caoyingjunz/gopixiu/api/server/router/demo"
-	"github.com/caoyingjunz/gopixiu/api/server/router/k8s"
 	"github.com/caoyingjunz/gopixiu/cmd/app/options"
 	"github.com/caoyingjunz/gopixiu/pkg/pixiu"
 )
@@ -71,9 +71,9 @@ func NewServerCommand() *cobra.Command {
 }
 
 func InitRouters(opt *options.Options) {
-	demo.NewRouter(opt.GinEngine) // 注册 demo 路由
-	cicd.NewRouter(opt.GinEngine) // 注册 cicd 路由
-	k8s.NewRouter(opt.GinEngine)  // 注册 k8s 路由
+	demo.NewRouter(opt.GinEngine)  // 注册 demo 路由
+	cicd.NewRouter(opt.GinEngine)  // 注册 cicd 路由
+	cloud.NewRouter(opt.GinEngine) // 注册 cloud 路由
 }
 
 func Run(opt *options.Options) error {
